optimize the shaper ~25% reduction

+++ dsp/wrShaper.h
@@ -4,9 +4,9 @@
 
 extern const float log_lut[]; // make LUT public to project
 
-float shaper_apply( float     input
-                  , uint8_t   zone  // [0:3]
-                  , float     coeff // (0,1)
+float shaper_apply( float input
+                  , int   zone  // [0..3]
+                  , float coeff // [0..1)
                   );
 float* shaper_apply_v( float  shape
                      , float* shape_audio
